services/api/pyproject.toml
[project]
name = "rampready-api"
version = "0.1.0"
description = "RampReady MVP backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "alembic>=1.13.0",
  "psycopg[binary]>=3.1.18",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt>=4.1.3",
  "httpx>=0.27.0",
  "redis>=5.0.4",
  "apscheduler>=3.10.4",
  "python-slugify>=8.0.4"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "pytest-asyncio>=0.23.6",
  "black>=24.4.2",
  "ruff>=0.4.4"
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
